Move join to util and replace twine
